Add file navigation leader
On my latest project with deeply nested files I often want to navigate
to file that is in the same directory, as the file I am currently
editing. %:h Gives me the directory, so then all I need to type is the
filename (or another :h if I want to navigate out one more directory).