Don't run tests in a separate pane
I haven't found running tests in a separate pane to be all that useful
lately. The pane is generally too small to what I need to see, and
closing other panes sometimes confuses vim-tmux-runner. I am going to
try out the default behavior for a while - running the tests in the
current pane.